one createCheckins components for explore page
